+++ src/snackbar-component.ts
@@ -4,6 +4,7 @@ export class SnackbarComponent extends HTMLElement {
     private settings: SnackbarNotification;
     constructor(snackbar: SnackbarNotification) {
         super();
+        console.log(snackbar);
         this.settings = snackbar;
         this.render();
     }
@@ -12,6 +13,7 @@ export class SnackbarComponent extends HTMLElement {
         const target = e.currentTarget as HTMLElement;
         const index = parseInt(target.dataset.index);
         this.settings.buttons[index].callback();
+        this.remove();
     };
 
     private handleCloseClickEvent: EventListener = () => {
@@ -43,11 +45,11 @@ export class SnackbarComponent extends HTMLElement {
                     button.innerText = this.settings.buttons[i].label;
                     button.dataset.index = `${i}`;
 
-                    for (let k = 0; k < this.settings.buttons[i].classes.length; k++) {
+                    for (let k = 0; k < this.settings.buttons[i]?.classes?.length; k++) {
                         button.classList.add(this.settings.buttons[i].classes[k]);
                     }
 
-                    if (this.settings.buttons[i].ariaLabel) {
+                    if (this.settings.buttons[i]?.ariaLabel) {
                         button.setAttribute("aria-label", this.settings.buttons[i].ariaLabel);
                     }
 
@@ -75,6 +77,8 @@ export class SnackbarComponent extends HTMLElement {
         if (this.settings.autofocus){
             const closeButton:HTMLButtonElement = this.querySelector(".js-snackbar-close");
             if (closeButton){
+                // @ts-ignore
+                document.activeElement.blur();
                 closeButton.focus();
             }
         }
@@ -83,6 +87,8 @@ export class SnackbarComponent extends HTMLElement {
                 if (this.settings.buttons[i].autofocus){
                     const button:HTMLButtonElement = this.querySelector(`button[data-index="${i}"]`);
                     if (button){
+                        // @ts-ignore
+                        document.activeElement.blur();
                         button.focus();
                         break;
                     }