+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- type declaration files
+- new `append()` function
+- alternate CDN version
+
+### Fixed
+
+- autofocus bug ([#19](https://github.com/codewithkyle/notifyjs/issues/19))
+- action button bug ([#20](https://github.com/codewithkyle/notifyjs/issues/20))
+
 ## [2.1.1] - 2021-01-21
 
 ### Fixed