+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0] - 2021-03-25
+
 ### Added
 
 - type declaration files
@@ -147,7 +149,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Entire existing codebase due to rewrite
 
-[unreleased]: https://github.com/codewithkyle/notifyjs/compare/v2.1.0...HEAD
+[unreleased]: https://github.com/codewithkyle/notifyjs/compare/v3.0.0...HEAD
+[3.0.0]: https://github.com/codewithkyle/notifyjs/compare/v2.1.0...v3.0.0
 [2.1.0]: https://github.com/codewithkyle/notifyjs/compare/v2.0.3...v2.1.0
 [2.0.3]: https://github.com/codewithkyle/notifyjs/compare/v2.0.1...v2.0.3
 [2.0.1]: https://github.com/codewithkyle/notifyjs/compare/v2.0.0...v2.0.1